requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
redis==5.0.1
hvac==1.1.1
docker==6.1.3
statistics==1.0.3.5
PyYAML==6.0.1
requests==2.31.0 